Move DottyTests into separate file.
